epoll: Ignore migrated tfds on checkpoint
In vz7-u6 the strategy for unknown target files has been
to ignore such targets, but in vz7-u7 we've introduced
additional kcmp syscall mode which would allow us to make
sure that the specified target is exactly the file struct
matched to the file descriptor number. Thus we started to
examinate every file descriptor to be sure the targets
are in proper state. Still there is a lack of support
that named migrated targets, ie when file is either
passed with scm message or inherited during the fork.
A child may open own epoll instance, put this inherited
target inside and close own tfd copy. We don't support
such scenarios yet (will be targeted in next release)
but people start complaining that this requirement is
too strong and cause more issues than solves. Thus lets
step back to the old strategy and ignore such targets.

https://jira.sw.ru/browse/PSBM-89961

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>