files: Select proper master fd when collecting fd
When choosing the master file which gonna be sending file
descriptor to the children we must not only look into
their PIDs but consider process tree relations, in particular
the child of a process might be choosen as a master and
epoll restore will fail because target files are simply
not present in child tree.

|  31964  31964  31964       epoll
|    585  31964  31964           epoll
|    586  31964  31964           epoll
|...
| (04.797121)    585: Error (criu/eventpoll.c:449): epoll: Can't add event on 0x000787: Bad file descriptor

That's because the target files are blong to 31964 and not
present in child 585, but because PID has wrapped the restore
of epoll started from 585 instead of 31964 and we have no
target files in the 585 because they all are inherited
from 31964.

Note the help of kcmp syscall which we started to use
since vz7-u8 won't helper here because on checkpoint
we choose the proper master and find 31964 carrying
all target files as expected.

https://jira.sw.ru/browse/PSBM-90293

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>