restore: pidfile -- Write from ve0 context
We've to jump into veX when restoring container simply because
we want restore procedure to be limited by the kernel (we have
a number of ve_is_super calls). Still the pidfile option might
point to a reference which doesn't belong the ve we restore
which may cause unexpected results on kernel changes. Thus
lets write pidfile a bit later when we've already jumped
back to ve0 environment.

https://jira.sw.ru/browse/PSBM-90289

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>