ve: tty -- Preallocate space for itty pool
To be able to allocate inheried terminal entries
we need to preallocate them in a root task thus
all children process would have access to this
shared memory. Previously they have been allocated
on demand, but such allocation may happen at the
moment when children which will access it is already
created, thus kernel won't propagate the memory area
bottom to top and such race window may cause error.

Instead we preallocate maximum possible number of entries
(which is number of processes plus number or teminals which
is usually a small number). Each entry is 32 byte in size
so should be acceptable tradeoff.

https://jira.sw.ru/browse/PSBM-90662

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>