epoll: Add suppport for migrated targets
Targets of epolls may belong to another process,
for example a child inherits epoll descriptor and
add own new file into it. For this sake we save
target's pid and fd in image then on restore we
arm a notifier which waits for target to appear,
fetches epoll descriptor from fdstore and adds
a target back.

https://jira.sw.ru/browse/PSBM-90012
Signed-off-by: Cyrill Gorcunov <gorcunov@gmail.com>

Feature: Epolls