files: make rst_remaps shmalloc'ed to see it's change in other process
In recent patch 9013d96f2943 ("files: Rework clean_one_remap to order
ghost dirs removal") the sort of these list was introduced. But the
change of the head of these list is not seen in other processes, thus
try_clean_remaps walks only random part of the list as it still points
to old first entry and now the old first entry is in the middle of the
list. These way leftover links created by criu remain after resume.

https://jira.sw.ru/browse/PSBM-93511
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: File remaps