net: add support for "net/core/*" sysctls
handle 'somaxconn' option.

https://jira.sw.ru/browse/PSBM-91415

Signed-off-by: Jan Dakinevich <jan.dakinevich@virtuozzo.com>

rebase to 3.15 mainstream:
Rename core_conf to vz_core_conf in image and change id 9 -> 1001.
Prepare for backward compatibility as now old images with core_conf(9)
will be read as new images with unix_conf(9) set to what was in
core_conf, so we need to fixup this.

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>