spfs: Retrieve NFS server IP address from mountinfo
Using hostname poses a lot of problems, including necessity to resolve
hostname on container with suspended networking and possibly incorrect
answer from nameserver due to loadbalancing or any other reason. We
should utilize IP address that is resolved and saved when connection
gets established - in can be retrieved from mountinfo.

https://jira.sw.ru/browse/PSBM-92070

Signed-off-by: Pavel Vokhmyanin <pvokhmyanin@virtuozzo.com>