mount/check-mounts: remove auxiliary cr-time mount from tree after umount
These mount is not mounted after mounts restore but it is in the mount
tree so then --check-mounts compares two trees they are different.

Also add bugons to check that this mount does not affect mount lists.
I think it's not a good idea to add a fake mount in random place in the
mount tree as these can affect mounts restore, deadlock or leave some
artifacts, but leave it there until we detect such a situation.

Also add mnt_slave list initialization, this was never used directly
before these patch.

Testing these in zdtm is a pain:
First criu should be compiled with CONFIG_BINFMT_MISC_VIRTUALIZED.
Second binfmt_misc should be virtualized and AFAICS it is done in VZ7
only yet (on ve not mntns).
So we will do a separate non-zdtm test in vzt.

https://jira.sw.ru/browse/PSBM-96082

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

rebase to 3.15 mainstream:
add mnt_unbindable to cr_time removal checks

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>