mount/restore: create auxiliary binfmt_misc mount in the root yard
The function add_cr_time_mount is used only twice, once on dump in
mount_and_collect_binfmt_misc and once on restore in populate_mnt_ns.

On dump the removed hunk did nothing as there is no root yard there.
On restore we have all mount trees of all mount namespaces merged
together. So on restore add_cr_time_mount added the auxiliary
binfmt_misc mount to the directory of the CT-root-mntns on the path
"./proc/sys/fs/binfmt_misc".

We don't actually need these as these mount can mess up with real
CT-mounts. So mount binfmt_misc as a direct child of the root yard on
the path "./binfmt_misc". Thus it will not interfere with CT mounts.

https://jira.sw.ru/browse/PSBM-96082
https://jira.sw.ru/browse/PSBM-96362

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>