mount: prohibit external mounts in non-root mntns
Currently when we add an external mount, it is identified by the path of
the mountpoint on dump irrelevant to mntns. If one creates a new mntns
in CT there could exist different mount with different superblock with
same path and detecting it as external is wrong.

https://jira.sw.ru/browse/PSBM-96385
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

rebase to mainstream 3.18:
split hunk from patch, and fix explanation
https://src.openvz.org/projects/OVZ/repos/criu/commits/034498b28

https://jira.sw.ru/browse/PSBM-144416
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Mounts