mount: restore shared options in clean mount for external mounts
After prohibiting to handle mounts in non-root CT mntnses as external
only one mount from these fs - "first" is restored as external and all
others are restored from the "first". In test we've had cgroup external
mount which propagated in multiple namespaces when mounting the "first"
mount, so we din't restore sharing on propagations right.

https://jira.sw.ru/browse/PSBM-97787

Fixes: fd48e2239 ("mount: fix external mounts so that they pass check-mounts")
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>