vz7: systemd-autofs-restart.sh: use exec3 and input from /dev/null
We've started to use vzctl exec instead of nsenter in the commit ("vz7:
systemd-autofs-restart.sh: use vzctl instead of nsenter for safety") and
now test detects situation that "vzctl exec mktemp -d" does not have
output though temporary directory is created successfuly, and thus
autofs services are not restarted.

The main idea (from @igor) is that we lose output because vzctl exec
calls intermediate bash and not the binary directly. Using exec3 we
skip intermediate bash and decrease posibility of failure.

Redirecting input of vzctl from /dev/null is generally good as vzctl and
it's descendants now can't read input of ancestors.

https://jira.sw.ru/browse/PSBM-98438

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>