vz7: don't get out VEX cgroup for iptables restore
We allow to do iptables-restore for pseudosuper, so we don't need to get
out to VE0 for it, see kernel patch: ("ve: use host ipt_mask for
pseudosuper in setsockopt and request module")

Also add a comment to join_ve0 about problems connected with it's
use, and how they can be fixed if needed in future. Now it is only used
once and the place is in host mntns, so we are fine.

https://jira.sw.ru/browse/PSBM-98702

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>