files: allow dumping open symlinks
To really open symlink file and not the regular file below it, one needs
to do open with O_PATH|O_NOFOLLOW flags. Looks like systemd started to
open /etc/localtime symlink this way sometimes, and before that nobody
actually used this and thus we never supported this in CRIU.

Error (criu/files-ext.c:96): Can't dump file 11 of that type [120777]
(unknown /etc/localtime)

Looks like it is quiet easy to support, as c/r of symlink file is almost
the same as c/r of regular one. We need to only make fstatat not
following links in check_path_remap.

https://jira.sw.ru/browse/PSBM-98888

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>