pie: restorer -- Show pids statistics on thread creation failure
When we're failing to restore threads (seems because of some kind
of race) there is no additional info present in the logs to figure
out what have happened.

Lets scan procfs and show all processes and threads, maybe it
reveal the race case.

https://jira.sw.ru/browse/PSBM-99230

NB: Probably we should zap this helper before the release.

Signed-off-by: Cyrill Gorcunov <gorcunov@virtuozzo.com>