Revert "./test/zdtm/static:conntracks: Support nftables"
This reverts commit cdbc841d41eb6ef53d79a7fe1cdd528f96fb6297.

It is been found that nft interface is yet unstable at least
in term of how we need to handle persistent data between c/r
cycle. In particular nft produce statistics "counters/bytes"
in output if been called without `-s` key.

Same time it is yet unclear how to handle conntracs in such
case.

After all we decide to zap this test for a while until
we figure out how to handle the output properly.

https://jira.sw.ru/browse/PSBM-99112

Signed-off-by: Cyrill Gorcunov <gorcunov@gmail.com>