mount: add external_slavery detection
If an external mount before checkpointing had an external master_id, it
should also have this after restoring and in contrary if it had internal
master_id it should be internal after restoring. Now we always assume
that if external mount has any master_id than these master_id is
external but it's not true.

For instance if we give an external private mount to a CT, the CT user
can make these mount shared, bind it somethere, and make these mount
slave. These way external mount can have internal master.

We already detect real state of internal/external master in
resolve_shared_mounts, so save it and use it in do_bind_mount.

Note 1: Sadly we already have internal_sharing variable and I understand
that I introduce some naming inconsistency with new external_slavery
name. Actually I like external_ more because mount_info is zero alloced,
and it is more convenient to have all mounts with no external
sharing/slavery by default, and only set it when we detect it.

Note 2: Sadly criu is not yet ready for a test in which we have external
mount being a slave of other mount in testct. But at least internal
slavery would not silently switch to external slavery after c/r.

https://jira.sw.ru/browse/PSBM-100080

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>