mount: make honest external/internal sharing detection
What we had before was wrong and only worked for auto detected external
mounts. The only right way to detect that the mount has external sharing,
is to check if there is a mount in criu's mount namespace with the same
shared id.

Partially revert hunk from fd48e2239fc3 ("mount: fix external mounts so
that they pass check-mounts")

https://jira.sw.ru/browse/PSBM-100080

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>