util: merge call_helper_process and call_in_child_process
They are doing almost the same thing but on restore we need to also lock
last_pid_mutex (if we have task_entries we are on restore).

More over on dump same as on restore we don't want to reuse proc self fd
cache from previous process in case of pid reuse.

https://jira.sw.ru/browse/PSBM-101289

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>