util: merge call_helper_process and call_in_child_process
They are doing almost the same thing but on restore we need to also lock
last_pid_mutex (if we have task_entries we are on restore).

More over on dump same as on restore we don't want to reuse proc self fd
cache from previous process in case of pid reuse.

https://jira.sw.ru/browse/PSBM-101289
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

rebase to mainstream 3.18:
merge fixup and rework to atomic
https://src.openvz.org/projects/OVZ/repos/criu/commits/4bbef43ee

https://jira.sw.ru/browse/PSBM-144416
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Call in helper process