util: don not use cached proc self fd in call_in_child_process
A helper child can open /proc/self/, if we cache it and the next helper
will get the same pid it would reuse this fd from already deleted
process. So lets don't use a cached fd if we are in child process.
Previousely we did close_pid_proc() as a cleanup after child finishes to
resolve the problem, but that approach does not work everywhere e.g.
when sfds_protected is set.

Add an atomic variable in_child_process it is zero in the beginning, if
call_in_child_process is used N times recursively, then in_child_process
would be N. There is also a posibility that call_in_child_process is
called from threads, in this case in_child_process would be a summ of
how many times was called from each thread, so it can be kinda
false-positive, as even if thread didn't call call_in_child_process,
some other thread could. We also use atomic variable exactly for this
multithreaded case as increments and decrements of in_child_process
which is shared between threads should be contention safe.

Feature: Call in helper process

https://jira.sw.ru/browse/PSBM-101289
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>