util: don not use cached proc self fd in call_in_child_process
A helper child can open /proc/self/, if we cache it and the next helper
will get the same pid it would reuse this fd from already deleted
process. So lets don't use a cached fd if we are in child process.
Previousely we did close_pid_proc() as a cleanup after child finishes to
resolve the problem, but that approach does not work everywhere e.g.
when sfds_protected is set.

https://jira.sw.ru/browse/PSBM-101289

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>