sk-unix: fix not correct assertion in on_stage_change() that fails restore stage
In on_stage_change() handler we should skip sockets that are not peers. In previous
version condition was not fully correct: we should also skip STREAM sockets in
ESTABLISHED state, this sockets also couldn't be peers.

Also, if we caught some inconsistency we shouldn't fail restore stage at all,
it seems sufficient to simply write about possible bug in logs.

https://jira.sw.ru/browse/PSBM-102972

Signed-off-by: Alexander Mikhalitsyn (Virtuozzo) <alexander@mihalicyn.com>