mount: replace CRTIME_MNT_ID with HELPER_MNT_ID
Root yard mount also has mnt_id == 0 so it will look better with a new name.

Also in near future we might want to create additional mount helpers to support
mounts in CT with no fsroot mounted.

Also on restore crtime mount is mounted as a child of the root yard so we don't
need to umount them. All helper mounts should be destroyed together with service
mount namespace.

https://jira.sw.ru/browse/PSBM-58617

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>