util: add get_relative_path helper
This is a smart way of getting relative paths:

1) Always returns relative path, no unexpected starting '/';
2) Detects subpath even if path formats are different, only real directory
names matter.

We have many places where we need to cut subpath from path. Different code
blocks doing this job spread widely across the codebase for instance see:
cut_root_for_bind and root_path_from_parent. But those implementations rely on
the fact that subpath's and path's formats are the same.

https://jira.sw.ru/browse/PSBM-58617

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

+++
util: make get_relative_path understand paths with '.'s

Sadly I've missed that ns_mountpoint on dump is prefixed with "./" and
when we use get_relative_path on it the "." between slashes is
considered as normal path component. So to support this kind of paths
lets skip "^./", "/./" and "/.$" at the time of slash skipping.

Now if you call get_relative_path("./a/b", "a/") you get "b" not "null"
as before the patch. Also any silly paths are supported too:

get_relative_path("./a////.///./b//././c", "///./a/b") == "c"

Note ".." in path is stil unsupported, but it would be to hard to
fix it.

https://jira.sw.ru/browse/PSBM-106869

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

rebase to 3.15 mainstream:
merge fixup

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>