mount: move set_is_overmounted before doing trees merge
Function mnt_is_overmounted is designed to detect if mount is overmounted in
current tree using comparison of mountpoints of neighbour mounts for detection.
We wan't to get actual overmounts in dumped tree, we don't expect that helper
mounts we add or merging will introduce new overmounts. So let's do overmount
detection earlier before adding helpers and doing merge

Also use ->ns_mountpoint in comparison as ->mountpoint can change. We plan to
get rid of ->mountpoint everywhere where we can use ->ns_mountpoint because of
it.

https://jira.sw.ru/browse/PSBM-58617

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>