check-mounts: pass for_dump false to collect_mntinfo
The meaning of check-mounts is to compare mount tree from images and the
mount tree created based on those images, no mounts should be skipped
implicitly, so set for_dump flag to false.

Also this for_dump flag is passed to fstype->parse helpers, and for nsfs
we can benefit from it to disable nsid lookup for check-mounts because
namespaces are not collected after restore and lookup will fail without
namespaces.

https://jira.sw.ru/browse/PSBM-102357

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>