cr-restore: move CR_STATE_RESTORE switch to restore_root_task
The main criu task switches to the CR_STATE_FORKING, but the next stage
CR_STATE_RESTORE is special and is switched to by the root task instead.
And the main task is waiting for CR_STATE_RESTORE to finish.

Change it so that the actual switch is done in the main task, and the
root task still waits all other tasks and calls it's stage ending
callbacks (ve_itty_resolve and ve_itty_resolve) and gives the power to
switch stage to the main task.

This way the main task would also be able to call it's own stage ending
callbacks for CR_STATE_FORKING, which is a real goal of this patch.

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Mounts