mount-v2: restore_mount_sharing_options at the end of forking stage
Function restore_mount_sharing_options should be called from service
mount namespace as it needs to have access to external mount source.
After forking stage the main criu task is the only one left in service
mount namespace. So if we want to delay it to the end of forking stage,
we need to call it from the main criu task.

Create special helper fini_restore_mntns_v2 for mounts-v2 actions which
need to be dalayed to the end of the forking stage.

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Mounts