mount-v2: create internal yard mounts for each mntns
We wan't to split mount restore into two steps: we will mount some part
of mounts to the tree in each mount namespace, but the rest of mounts
will be mounted later (mounts which depend from pid namespaces can be
mounted only after forking all tasks).

To be able to mount the rest of mounts we need to somehow handle if
their mountpoints in tree are already overmounted. For that we have
mp_fd_id and mnt_fd_id for all mounted mounts. But it's not enough to be
able to mount through them, we also need a host procfs visible from "/"
in each mount namespace so that any task can mount through
/proc/self/fd/<id> files.

Also we would need to create some helper mounts on the first step to
make it easier to create the rest of mounts.

So we will put all those helper mounts in special place "internal root
yard" in each mount namespace. Let's put it in temporary directory
"/internal-yard-XXXXXX" as children mount to the root mount of mountns.
Put mountpoint path on shared memory as we will resolve XXXXXX's real
value only when do mkdtemp from root task and it should be also visible
to main criu task.

Note: This internal yard is not in the tree and mntinfo, it will be
added to the tree only temporary for mounting. We will umounted it then
done with mounts restore. So check-mounts still works as before.

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>