mount-v2: temporary mount enabled internal yards to mntns trees
To mount them we simply add them to the tree before general mounting
code. We mount internal yard to "/internal-yard-XXXXXX" in the root
mount of the mountns.

We also umount internal yards in fini_restore_mntns_v2 when we don't
need them anymore with all descendants as they should not be visible
after restore.

We also need to remove temporary directory. We need to enter owner
userns of the mntns to have all permissions to remove the temporary dir
of internal yard. Also add cleanup_internal_yards helper to cleanup it
on error paths.

Untill now there is no actuall mounting as yards are not enabled, but
when enabled they will be created. We wan't to enable internal yards
only in mount namespaces which actually need to use it.

v2: disable internal yards by default

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>