mount-v2: add resolve_fd_path helper
The meaning of this function is to get an fd to a dentry at some path
from the given mount. It uses mnt_fd_id and mp_fd_id-s of already
mounted mounts and thus handles any type of overmounts which can hide
the path from normal filesystem view.

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Mounts