sk-unix: make bindmounted unix sockets work on mounts-v2
The problem was that on mounts v2 we mount all mounts
in plain structure (to the same directory), when
we restore bindmounted unix socket we should create
and bind socket before bindmounting it. But on this stage
mount v2 engine is not yet moved mounts to the tree structure.
To get around this problem we make temporary fake tree, move
needed mount to that tree, bind socket and then move mount
to the plain structure in the unix_prepare_bindmount().

https://jira.sw.ru/browse/PSBM-104800
Signed-off-by: Alexander Mikhalitsyn <alexander.mikhalitsyn@virtuozzo.com>

rebase to mainstream 3.18:
merge fixup:
https://src.openvz.org/projects/OVZ/repos/criu/commits/4dafb92f6

https://jira.sw.ru/browse/PSBM-144416
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Unix sockets