files-reg: fix open_remap_spfs_linked remap has no mnt_id
We've got "Error (criu/files-reg.c:888): The -1 mount is not found for
ghost" because create_link_remap dumps RegFileEntry without mnt_id for
link-remap, that is because same mnt_id is assumed as for original rfi
from which we did the remap.

So just remove rmi part to return to old behaviour, we can do it as spfs
link remaps are remaped only once, see dump_linked_remap_type.

https://jira.sw.ru/browse/PSBM-105661

Fixes: bc1261852 ("files-reg: teach create_spfs and
open_remap_spfs_linked to work with mounts")
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>