zdtm: optionally skip inode number and fhandle in fanotify_obj_cmp
The problem with fanotify_obj_cmp is that
Checkpoint/restore for filesystem objects results in creation
of inodes from scratch. In overlayfs inodes are created on the fly,
and act as proxies to the underlying filesystem's inodes.
For such cases c/r will result in varying inode numbers in fdinfo
of fanotify file descriptor. In fanotify_obj_cmp we no not want
to check that inode numbers are matching for such filesystems.
Additional flag for this function allows to skip this check.

https://jira.sw.ru/browse/PSBM-104961

Signed-off-by: Valeriy Vdovin <valeriy.vdovin@virtuozzo.com>