mount: restore sysfs from proper netns
Sysfs mounts save current netns when mounting and show different set of
network devices in /sys/class/net based on it.

So let's use the KERNFS_GET_NS ioctl interface to dump it, so that on
restore we can call mount syscall from the right netns.

In case of no ioctl or migration from old image we would assume sysfs is
owned by root netns.

https://jira.sw.ru/browse/PSBM-105161

v2: error to warning in case of no KERNFS_GET_NS ioctl

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

rebase to 3.15 mainstream:
don't search sysfs owner netns if we don't dump mounts to fix zdtm,
switch to after_parse callback which we already used for nsfs data
parsing to be able to use open_mountpoint

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>