zdtm: added overlayfs setup routine to zdtm lib
With growing number of overlayfs code, there is a need to repeat the same setup
routine for overlayfs. There is a list of directories that need to be set up
and created before calling mount. 'overlayfs_setup' is added to zdtm lib to
perform some of the most generic steps like making directories and forming
mount options string.

https://jira.sw.ru/browse/PSBM-104961

Signed-off-by: Valeriy Vdovin <valeriy.vdovin@virtuozzo.com>