mount: replace sysfs_parse with dump_sysfs
Sadly we can't use open_moutpoint in ->parse callbacks as we call them
before fully constructing mount tree.

  (00.009508) Error (criu/mount.c:1371): mnt: The file system 0x45 0x45
  (0x46) sysfs ./zdtm/static/mntns_overmount.test/1 is inaccessible
  ####### Test zdtm/static/mntns_overmount FAIL at CRIU dump #######

https://ci.openvz.org/job/CRIU/job/CRIU-virtuozzo-stable/job/vz7-u15/214

So we need to move the code which gets netns of sysfs to ->dump
callback. This way we also need search_nested_netns_sysfs function to
copy network namespace id of sysfs mount to all it's bind-mounts.

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>