zdtm: put common prepare_dirname function to library
It is a useful helper to make all newly created files, directories and
mounts in dirname automatically be umounted or removed when test
finishes, so that we do not need to manually cleanup them. There will be
also no unexpected propagation from dirname as it is private. Only
restriction is that the test should be run in separate mount namespace,
which is destroyed after test finishes, so this is suitable to ns and
uns flavours only.

While on it also separate user and system headers with empty line
according to CRIU coding style.

https://jira.sw.ru/browse/PSBM-104871
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: NSFS bindmounts