vz: use 1000+x field ids for vz-only field ids
These fields are related to features which are not so easy to be ported,
so we must secure their ids from collisions with mainstream criu on
rebase.

Note: This patch breaks image format compatibility for all previous
vz7-u15 criu builds!!! But those builds were not released and should not
be used anywhere exept for tests, so everything should be fine.

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>