util: make get_relative_path understand paths with '.'s
Sadly I've missed that ns_mountpoint on dump is prefixed with "./" and
when we use get_relative_path on it the "." between slashes is
considered as normal path component. So to support this kind of paths
lets skip "^./", "/./" and "/.$" at the time of slash skipping.

Now if you call get_relative_path("./a/b", "a/") you get "b" not "null"
as before the patch. Also any silly paths are supported too:

get_relative_path("./a////.///./b//././c", "///./a/b") == "c"

Note ".." in path is stil unsupported, but it would be to hard to
fix it.

Fixes: cc71dc93d ("util: add get_relative_path helper")

https://jira.sw.ru/browse/PSBM-106869

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>