cgroup: add some additional debugging information
Previousely there were no way to see if we actually walking some
directory in some cgroup hierarchy on dump => add it.

When we rewrite cgroup roots, lets at least print first cgroup hierarchy
name from this controller, so that we know for which controller the
rewrite takes place.

Curently there is always cname[0] as it is originated from
new_controller, add a bug to check it.

v2: bug on zero n_cname

Reviewed-by: Alexander Mikhalitsyn <alexander.mikhalitsyn@virtuozzo.com>
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

Feature: Debug