mounts-v2: add new criu optioin --mounts-compat and drop --mounts-v2
Now we are enabling new mounts engine by default in criu. If one wants
to use old mounts engine there is an option "--mounts-compat" for it,
which has the oposite meaning.

We can safely drop --mounts-v2 and also reuse rpc image id as it was
only used in tests and we will update tests. Now if this option is
passed to criu it will fail.

Updating zdtm and jenkins scripts with a new option. We still be testing
"--mounts-compat" in zdtm separately.

Note "--mounts-compat" is prepended in zdtm so that "--no-mounts-compat"
from test description can overwrite it.

https://jira.sw.ru/browse/PSBM-106066

Reviewed-by: Alexander Mikhalitsyn <alexander.mikhalitsyn@virtuozzo.com>
Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>