seize: recursively thaw freezers in freeze_processes
From libvzctl commit 7186bdba we started to set
freezer.state recursively. But looks like we need
to do that recursively only on container stop
and only to thaw tasks. Side effect of this change
is that we freeze tasks recursively and loose
user freezers state as consequence. This is not
critical for now because we not support this in CRIU
but we should return to this once time permit.

Fixes https://jira.sw.ru/browse/PSBM-107590

Signed-off-by: Alexander Mikhalitsyn <alexander.mikhalitsyn@virtuozzo.com>
Suggested-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>
Reviewed-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>

rebase to 3.15 mainstream:
a) rename set_freezer_state_recurse to freezer_write_state_recurse
to be consistent with ms code
b) reuse freezer_open and freezer_write_state from ms code

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>