freezer: added cmd-line option to skip 'freezer.state' dump/restore
During 'vzctl snapshot CT' command execution vzctl first freezes
container, then spawnes criu with 'dump' command. Because container
is in FROZEN state, criu needs --freeze-cgroup flag to call
freeze_processes, which will unfreeze it back to be able seize to
container processes. The same --freeze-cgroup flag is later used
to determine if freezer state should be dumped to image.
In case of snapshot it is impossible to dump orignal state of root
container's freeze cgroup due to the fact that it has already been
overwritten by vzctl and it's in FROZEN state.
Because of that at snapshot-switch this FROZEN state will be restored
by criu and will lead to hang of any processess joining that container.
This problem is new because at snapshot-switch criu did not restore cgroup
properties by default and thus "FROZEN" freezer.state in dump image was
ignored. But due to https://jira.sw.ru/browse/PSBM-106413 criu now
restores properties and so it has to know weather or not it should
handle freezer state. For common case we could only add flag to not dump
freezer state during criu dump. But for compatibility of older snapshots
that dumped FROZEN states before this fix we also need flag at criu
restore that tells criu to ignore freezer.state property from dump
image.

https://jira.sw.ru/browse/PSBM-107590
Signed-off-by: Valeriy Vdovin <valeriy.vdovin@virtuozzo.com>